backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.28.1
lxml==5.3.1
lxml_html_clean==0.4.2
feedparser==6.0.11
newspaper3k==0.2.8
openai==1.61.1
